chore: restructure shared and react to use /core package
