+++ wallets/core/package.json
@@ -3,13 +3,17 @@
   "version": "0.37.0",
   "license": "MIT",
   "type": "module",
-  "source": "./src/legacy/index.ts",
-  "main": "./dist/index.js",
-  "typings": "./dist/legacy/index.d.ts",
+  "source": "./src/mod.ts",
+  "main": "./dist/mod.js",
+  "typings": "./dist/mod.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/legacy/index.d.ts",
-      "default": "./dist/index.js"
+      "types": "./dist/mod.d.ts",
+      "default": "./dist/mod.js"
+    },
+    "./legacy": {
+      "types": "./dist/legacy/mod.d.ts",
+      "default": "./dist/legacy/mod.js"
     }
   },
   "files": [
@@ -17,7 +21,7 @@
     "src"
   ],
   "scripts": {
-    "build": "node ../../scripts/build/command.mjs --path wallets/core --inputs src/legacy/index.ts",
+    "build": "node ../../scripts/build/command.mjs --path wallets/core --inputs src/mod.ts,src/legacy/mod.ts",
     "ts-check": "tsc --declaration --emitDeclarationOnly -p ./tsconfig.json",
     "clean": "rimraf dist",
     "format": "prettier --write '{.,src}/**/*.{ts,tsx}'",
@@ -29,7 +33,6 @@
     "react-dom": "^17.0.0 || ^18.0.0"
   },
   "dependencies": {
-    "@rango-dev/wallets-shared": "^0.36.0",
     "rango-types": "^0.1.69"
   },
   "publishConfig": {
