+++ wallets/core/src/legacy/helpers.ts
@@ -1,5 +1,8 @@
+import type { Network } from './types.js';
 import type { Options } from './wallet.js';
-import type { Network } from '@rango-dev/wallets-shared';
+import type { BlockchainMeta } from 'rango-types';
+
+import { Networks } from './types.js';
 
 export function formatAddressWithNetwork(
   address: string,
@@ -37,3 +40,36 @@ export function needsCheckInstallation(options: Options) {
   const { checkInstallation = true } = options.config;
   return checkInstallation;
 }
+
+export const getBlockChainNameFromId = (
+  chainId: string | number,
+  blockchains: BlockchainMeta[]
+): Network | null => {
+  chainId =
+    typeof chainId === 'string' && chainId.startsWith('0x')
+      ? parseInt(chainId)
+      : chainId;
+
+  /*
+   * Sometimes providers are passing `Network` as chainId.
+   * If chainId is a `Network`, we return itself.
+   */
+  const allNetworks = Object.values(Networks);
+  if (allNetworks.includes(String(chainId) as Networks)) {
+    return chainId as Networks;
+  }
+
+  if (chainId === 'Binance-Chain-Tigris') {
+    return Networks.BINANCE;
+  }
+  return (
+    blockchains
+      .filter((blockchainMeta) => !!blockchainMeta.chainId)
+      .find((blockchainMeta) => {
+        const blockchainChainId = blockchainMeta.chainId?.startsWith('0x')
+          ? parseInt(blockchainMeta.chainId)
+          : blockchainMeta.chainId;
+        return blockchainChainId == chainId;
+      })?.name || null
+  );
+};
