--- wallets/core/src/legacy/index.ts
@@ -1,6 +1,0 @@
-export type { EventHandler, State, Options } from './wallet.js';
-export * from './types.js';
-
-export { Persistor } from './persistor.js';
-export * from './helpers.js';
-export { default } from './wallet.js';
