+++ wallets/core/src/legacy/types.ts
@@ -1,11 +1,107 @@
 import type { State as WalletState } from './wallet.js';
-import type {
-  NamespaceData,
-  Network,
-  WalletInfo,
-  WalletType,
-} from '@rango-dev/wallets-shared';
 import type { BlockchainMeta, SignerFactory } from 'rango-types';
+
+export enum Networks {
+  BTC = 'BTC',
+  BSC = 'BSC',
+  LTC = 'LTC',
+  THORCHAIN = 'THOR',
+  BCH = 'BCH',
+  BINANCE = 'BNB',
+  ETHEREUM = 'ETH',
+  POLYGON = 'POLYGON',
+  TERRA = 'TERRA',
+  POLKADOT = '',
+  TRON = 'TRON',
+  DOGE = 'DOGE',
+  HARMONY = 'HARMONY',
+  AVAX_CCHAIN = 'AVAX_CCHAIN',
+  FANTOM = 'FANTOM',
+  MOONBEAM = 'MOONBEAM',
+  ARBITRUM = 'ARBITRUM',
+  BOBA = 'BOBA',
+  OPTIMISM = 'OPTIMISM',
+  FUSE = 'FUSE',
+  CRONOS = 'CRONOS',
+  SOLANA = 'SOLANA',
+  MOONRIVER = 'MOONRIVER',
+  GNOSIS = 'GNOSIS',
+  COSMOS = 'COSMOS',
+  OSMOSIS = 'OSMOSIS',
+  AXELAR = 'AXELAR',
+  MARS = 'MARS',
+  STRIDE = 'STRIDE',
+  MAYA = 'MAYA',
+  AKASH = 'AKASH',
+  IRIS = 'IRIS',
+  PERSISTENCE = 'PERSISTENCE',
+  SENTINEL = 'SENTINEL',
+  REGEN = 'REGEN',
+  CRYPTO_ORG = 'CRYPTO_ORG',
+  SIF = 'SIF',
+  CHIHUAHUA = 'CHIHUAHUA',
+  JUNO = 'JUNO',
+  KUJIRA = 'KUJIRA',
+  STARNAME = 'STARNAME',
+  COMDEX = 'COMDEX',
+  STARGAZE = 'STARGAZE',
+  DESMOS = 'DESMOS',
+  BITCANNA = 'BITCANNA',
+  SECRET = 'SECRET',
+  INJECTIVE = 'INJECTIVE',
+  LUMNETWORK = 'LUMNETWORK',
+  BANDCHAIN = 'BANDCHAIN',
+  EMONEY = 'EMONEY',
+  BITSONG = 'BITSONG',
+  KI = 'KI',
+  MEDIBLOC = 'MEDIBLOC',
+  KONSTELLATION = 'KONSTELLATION',
+  UMEE = 'UMEE',
+  STARKNET = 'STARKNET',
+  TON = 'TON',
+
+  // Using instead of null
+  Unknown = 'Unkown',
+}
+
+export enum Namespace {
+  Solana = 'Solana',
+  Evm = 'EVM',
+  Cosmos = 'Cosmos',
+  Utxo = 'UTXO',
+  Starknet = 'Starknet',
+  Tron = 'Tron',
+}
+
+export type NamespaceData = {
+  namespace: Namespace;
+  derivationPath?: string;
+};
+
+export type WalletType = string;
+export type Network = string;
+
+export type InstallObjects = {
+  CHROME?: string;
+  FIREFOX?: string;
+  EDGE?: string;
+  BRAVE?: string;
+  DEFAULT: string;
+};
+
+export type WalletInfo = {
+  name: string;
+  img: string;
+  installLink: InstallObjects | string;
+  color: string;
+  supportedChains: BlockchainMeta[];
+  showOnMobile?: boolean;
+  isContractWallet?: boolean;
+  mobileWallet?: boolean;
+  namespaces?: Namespace[];
+  singleNamespace?: boolean;
+  needsDerivationPath?: boolean;
+};
 
 export type State = {
   [key: string]: WalletState | undefined;
