+++ wallets/core/src/mod.ts
@@ -1,0 +1,1 @@
+// This file will be updated to export Hub methods and types.
