+++ wallets/react/src/index.ts
@@ -2,6 +2,3 @@ export * from './legacy/helpers.js';
 export { default as Provider } from './provider.js';
 export { useWallets } from './legacy/hooks.js';
 export * from './legacy/types.js';
-
-export type { EventHandler } from '@rango-dev/wallets-core';
-export { Events, readAccountAddress } from '@rango-dev/wallets-core';
