+++ wallets/shared/package.json
@@ -27,6 +27,7 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
+    "@rango-dev/wallets-core": "0.37.0",
     "ethers": "^5.7.2",
     "rango-types": "^0.1.69"
   },
